requirements.txt
flask
pandas
chromadb
sentence-transformers
openai
python-dotenv
tiktoken
scikit-learn
pyarrow
numpy